update core behavior
Specifically:

1) First try to load the string using standard json module before
applying the scrubbing heuristics

2) expose an arugment for allowing yaml loading-- I've decided to keep
it in here because yaml *is* a superset of json, but its dangerous to
keep the yaml.load() condition in by default because it will load any
yaml that might be present in the content.. which might not be desired.

Overall, this is still a hacky project that could use some improvement.